requirements.txt
Django==4.1.7
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-configurations==2.4.1
dj-database-url==1.0.0
django[argon2]
pymemcache==4.0.0
django-debug-toolbar
django-registration
django-allauth
djangorestframework
drf_yasg
django-filter
djangorestframework-simplejwt[crypto]
PyJWT
django-versatileimagefield
